libibverbs: Add cross-channel QP flags
The cross-channel feature allows to execute WQEs that involve
synchronization of IO operations’ on different QPs.

This capability enables to program complex flows with a single
function call, hereby significantly reducing overhead associated
with IO processing.

The queue pairs can be configured to work as a “sync master queue”
or “sync slave queues”.

QP property flags to indicate if queues are slave or master:
  * IB_QP_CREATE_MANAGED_SEND indicates that posted send work requests
    will not be executed immediately and requires enabling.

  * IB_QP_CREATE_MANAGED_RECV indicates that posted receive work
    requests will not be executed immediately and requires enabling.

  * IB_QP_CREATE_CROSS_CHANNEL declares the QP to work in cross-channel
    mode. If IB_QP_CREATE_MANAGED_SEND and IB_QP_CREATE_MANAGED_RECV are
    not provided, this QP will be sync master queue, else it will be sync
    slave.

  * IBV_QP_CREATE_IGNORE_SQ_OVERFLOW configures QP to discard overflow
    indications on send queue.

  * IBV_QP_CREATE_IGNORE_RQ_OVERFLOW instructs QP to discard overflow
    indications on receive queue.

Change-Id: I11ae5161a4e65d1bea1fc36358f1a2f2825d53e7
Signed-off-by: Leon Romanovsky <leonro@mellanox.com>